feat(toolbar): add CanvasToolbar component for action management
Introduced a new CanvasToolbar component to enhance the workflow management experience. This toolbar categorizes actions, allows for efficient searching, and integrates with the existing action management system. The implementation includes improved styling and layout for better user interaction.
